requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
PyGithub==2.1.1
httpx==0.26.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bcrypt==4.1.2
